+++ src/LASRstages/ivf.h
@@ -8,7 +8,7 @@ class LASRivf: public Stage
 public:
   LASRivf() = default;
   bool process(PointCloud*& las) override;
-  double need_buffer() const override { return res; };
+  double need_buffer() const override { return std::max(res_x, res_y); };
   bool set_parameters(const nlohmann::json&) override;
   std::string get_name() const override { return "ivf"; };
 
@@ -16,7 +16,7 @@ public:
   LASRivf* clone() const override { return new LASRivf(*this); };
 
 private:
-  double res;
+  double res_x,res_y,res_z;
   int n;
   int classification;
   bool force_map;